rename & move widget
