+++ libs/media/src/lib/file/upload-widget/upload-widget.component.html
